more refactoring and cleanup
