+++ src/gui/mod.rs
@@ -1,3 +1,4 @@
+pub mod draw_creation_gui_sys;
 pub mod graphs;
 pub mod gui_systems;
 pub mod imgui_wrapper;
